Namespace/untangle CSS classes using BEM
The CSS was becoming a bit messy, with classes such as “light” or
“menu”, making it more likely to accidentally use a class meant for one
page in another page, making it harder to search for, and so on. I chose
to standardise using the BEM (block-element-modifier) convention; see
http://getbem.com/naming/.

Again, I prefixed everything with `vm-`, in this case to make explicit
which styles are *ours*, as opposed to from Foundation, or from other
sources (e.g. Facebook and Twitter button styles).

Additionally, the CSS was nested, which makes it harder to figure out
exactly which CSS rules are being applied, as that suddenly becomes
context-dependent. Also, the rules for CSS specificity become more
complicated with nesting going on (see 
https://developer.mozilla.org/en-US/docs/Web/CSS/Specificity). So I made
all our CSS flat.

In a later commit I will restructure *where* there styles go exactly, as
that is still a bit messy. I deliberately didn’t do that in this commit
so it’s easier to review.

Test plan: since I could have made typos during this process, and
because specificity of rules changes when you flatten CSS, I carefully
tested the styles before and after, and made some specificity fixes. I
didn’t test across different browsers, as I’m not introducing new
behaviour here, and specificity rules are very consistent across
browsers (luckily!). I did test with different browser sizes, to test
the breakpoints.

The “join” page changes slightly as a result of this change, but I think
that’s acceptable, as I think the previous behaviour was a bug (the
`width: 90%` was applied multiple times because there were multiple
nested `<span>` elements).

One of the “faq” items is also slightly different to avoid an `<ol>`;
instead I made it consistent with other lists by using `&ndash;`.